WTE-105 Refactor fetchFeed action for Healthy Feed and Popular Feed.
